requirements.txt
# Core dependencies
selenium>=4.0.0
tqdm>=4.65.0
openai>=1.0.0
rich>=13.0.0

# Development dependencies
pylint>=3.0.0
flake8>=7.0.0
ruff>=0.2.0setuptools
anthropic
google-generativeai
